+++ src/Control/Monad/Trans/Fail.hs
@@ -51,6 +51,7 @@ import Control.Monad.Catch (MonadThrow (throwM))
 import Control.Monad.Cont
 import Control.Monad.Except
 import qualified Control.Monad.Fail as F
+import Control.Monad.RWS.Class (MonadRWS)
 import Control.Monad.Reader
 import Control.Monad.State
 import Control.Monad.Writer
@@ -67,6 +68,14 @@ import GHC.Stack
 #if MIN_VERSION_base(4,12,0)
 import Data.Functor.Contravariant
 #endif
+
+
+#if !(MIN_VERSION_base(4,13,0))
+#define IS_MONAD_STRING IsString e,
+#else
+#define IS_MONAD_STRING
+#endif
+
 
 -- | `FailT` transformer with `Identity` as the base monad.
 type Fail e = FailT e Identity
@@ -266,7 +275,7 @@ instance Functor m => Functor (FailT e m) where
   fmap f (FailT m) = FailT (fmap (fmap f) m)
   {-# INLINE fmap #-}
 
-instance Monad m => Applicative (FailT e m) where
+instance (IS_MONAD_STRING Monad m) => Applicative (FailT e m) where
   pure = FailT . pure . Right
   {-# INLINE pure #-}
 
@@ -283,7 +292,7 @@ instance Monad m => Applicative (FailT e m) where
   {-# INLINE (*>) #-}
 
 -- | Short-circuites on the first failing operation.
-instance (IsString e, Monad m) => Monad (FailT e m) where
+instance (IS_MONAD_STRING Monad m) => Monad (FailT e m) where
   FailT m >>= k =
     FailT $
       m >>= \case
@@ -309,7 +318,7 @@ instance Traversable f => Traversable (FailT e f) where
   {-# INLINE traverse #-}
 
 -- | Short-circuits on the first successful operation, combines failures otherwise.
-instance Monad m => Alternative (FailT e m) where
+instance (IS_MONAD_STRING Monad m) => Alternative (FailT e m) where
   empty = FailT $ pure (Left [])
   {-# INLINE empty #-}
   FailT m <|> FailT k = FailT $ do
@@ -323,7 +332,7 @@ instance Monad m => Alternative (FailT e m) where
 
 -- | Executes all monadic actions and combines all successful results using a `Semi.Semigroup`
 -- instance. Combines together all failures as well, until a successful operation.
-instance (Monad m, Semi.Semigroup a) => Semi.Semigroup (FailT e m a) where
+instance (IS_MONAD_STRING Monad m, Semi.Semigroup a) => Semi.Semigroup (FailT e m a) where
   (<>) (FailT m) (FailT k) = FailT $ do
     mres <- m
     kres <- k
@@ -338,14 +347,14 @@ instance (Monad m, Semi.Semigroup a) => Semi.Semigroup (FailT e m a) where
           Right y -> pure $ Right (x Semi.<> y)
   {-# INLINEABLE (<>) #-}
 
-instance (Monad m, Semi.Semigroup a) => Monoid (FailT e m a) where
+instance (IS_MONAD_STRING Monad m, Semi.Semigroup a) => Monoid (FailT e m a) where
   mempty = empty
   {-# INLINE mempty #-}
 #if !(MIN_VERSION_base(4,11,0))
   mappend = (Semi.<>)
 #endif
 
-instance (IsString e, MonadIO m) => MonadIO (FailT e m) where
+instance (IS_MONAD_STRING MonadIO m) => MonadIO (FailT e m) where
   liftIO = lift . liftIO
   {-# INLINE liftIO #-}
 
@@ -353,7 +362,7 @@ instance MonadTrans (FailT e) where
   lift = FailT . fmap Right
   {-# INLINE lift #-}
 
-instance (IsString e, MonadZip m) => MonadZip (FailT e m) where
+instance (IS_MONAD_STRING MonadZip m) => MonadZip (FailT e m) where
   mzipWith f (FailT a) (FailT b) = FailT $ mzipWith (liftA2 f) a b
   {-# INLINE mzipWith #-}
 
@@ -401,7 +410,7 @@ instance (Read e, Read1 m, Read a) => Read (FailT e m a) where
 instance (Show e, Show1 m, Show a) => Show (FailT e m a) where
   showsPrec = showsPrec1
 
-instance (IsString e, MonadReader r m) => MonadReader r (FailT e m) where
+instance (IS_MONAD_STRING MonadReader r m) => MonadReader r (FailT e m) where
   ask = lift ask
   {-# INLINE ask #-}
   local = mapFailT . local
@@ -409,7 +418,7 @@ instance (IsString e, MonadReader r m) => MonadReader r (FailT e m) where
   reader = lift . reader
   {-# INLINE reader #-}
 
-instance (IsString e, MonadState s m) => MonadState s (FailT e m) where
+instance (IS_MONAD_STRING MonadState s m) => MonadState s (FailT e m) where
   get = lift get
   {-# INLINE get #-}
   put = lift . put
@@ -417,13 +426,13 @@ instance (IsString e, MonadState s m) => MonadState s (FailT e m) where
   state = lift . state
   {-# INLINE state #-}
 
-instance (IsString e, MonadError e m) => MonadError e (FailT e m) where
+instance (IS_MONAD_STRING MonadError e m) => MonadError e (FailT e m) where
   throwError = lift . throwError
   {-# INLINE throwError #-}
   catchError = liftCatch catchError
   {-# INLINE catchError #-}
 
-instance (IsString e, MonadWriter w m) => MonadWriter w (FailT e m) where
+instance (IS_MONAD_STRING MonadWriter w m) => MonadWriter w (FailT e m) where
   writer = lift . writer
   {-# INLINE writer #-}
   tell = lift . tell
@@ -433,7 +442,10 @@ instance (IsString e, MonadWriter w m) => MonadWriter w (FailT e m) where
   pass = liftPass pass
   {-# INLINE pass #-}
 
-instance (IsString e, MonadCont m) => MonadCont (FailT e m) where
+-- | @since 0.1.1
+instance (IS_MONAD_STRING MonadRWS r w s m) => MonadRWS r w s (FailT e m)
+
+instance (IS_MONAD_STRING MonadCont m) => MonadCont (FailT e m) where
   callCC = liftCallCC callCC
   {-# INLINE callCC #-}