+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for `FailT`
 
+## 0.1.1.0
+
+* Remove unnecessary `IsString` constraint from instance definitions for GHC >= 8.10
+* Add `MonadRWS` instance.
+
 ## 0.1.0.0
 
 * Initial release